+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+vNext
+  * Add `Model.findAndCountAll()` support (thanks to @TerryMooreII)
+
 v0.9.1 - 3aeaa05 - Sep 21st 2017
   * Add `authenticate()` to `sequelize` which always resolves (thanks to @vFederer)
   * Fix a few documentation issues