+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 vNext
-  * Add `Model.findAndCountAll()` support (thanks to @TerryMooreII)
+  * Add support for `Model.findAndCount()` and it's alias `Model.findAndCountAll()` (thanks to @TerryMooreII)
 
 v0.9.1 - 3aeaa05 - Sep 21st 2017
   * Add `authenticate()` to `sequelize` which always resolves (thanks to @vFederer)